query/requirements.txt
grpcio-tools

